update save data facebook tich xanh

+++ src/Http/Controllers/FacebookController.php
@@ -6,10 +6,8 @@
 
 namespace TinhPHP\Tool\Http\Controllers;
 
-use Illuminate\Session\Store;
-use Illuminate\Support\Facades\Cookie;
-use Illuminate\Support\Facades\Log;
-use Illuminate\Support\Facades\Storage;
+use App\Services\TelegramService;
+use Illuminate\Support\Facades\Validator;
 use Intervention\Image\ImageManagerStatic as Image;
 
 /**
@@ -56,29 +54,59 @@ final class FacebookController extends ToolController
 
     public function avatar()
     {
+        $errors = [];
+
+        $fileNameUnique = md5(request()->header('User-Agent')) . '.png';
+
+        $avatar = '';
+
+        if (request()->has('submit')) {
+            $rules = [
+                'avatar' => 'required|file|mimes:jpeg,jpg,png|max:5120'
+            ];
+            $params = request()->only('avatar');
+
+            $validator = Validator::make($params, $rules);
+
+            if ($validator->fails()) {
+                $errors = $validator->messages()->toArray();
+            }
+
+            if (!$validator->fails()) {
+                // open an image file
+                $path = request()->file('avatar')->store('public/upload/tool/facebook-avatar/' . date('Y/m/d'));
+
+                $img = Image::make(storage_path('app/' . $path));
+
+                // resize image instance
+                $img->resize(500, 500);
+
+                // insert a watermark
+                $img->insert(public_path('site/img/tich-xanh-iframe.png'), 'center');
+
+                // save image in desired format
+                $fileName = 'tool/facebook-avatar/' . date('Y/m/d/') . $fileNameUnique;
+                $img->save(storage_path('app/public/upload/' . $fileName))->mime();
+                $avatar = asset('storage/upload/' . $fileName);
+
+                // push telegram
+                $text[] = '<strong>Tạo ảnh facebook tích xanh</strong>';
+                $text[] = '- Image: ' . $avatar;
+                TelegramService::send($text);
+
+                unlink(storage_path('app/' . $path));
+            }
+        }
+
         $data = [
+            'keyword' => 'Tạo ảnh đại diện Facebook, Tạo ảnh online , Tạo avatar , Facebook tích xanh, Tạo avatar tích xanh',
+            'description' => 'Tổng hợp full bộ icon facebook đầy đủ nhất, mới nhất với nhiều trạng thái khác nhau, công cụ viết status facebook kèm icon tiện lợi nhất chỉ cần click vào biểu tượng cảm xúc facebook và chọn nhiều icon facebook khác nhau',
+            'title' => 'Tạo ảnh đại diện Facebook có dấu tích xanh',
             'active_menu' => 'facebook_avatar',
+            'avatar' => $avatar,
+            'errors' => $errors
         ];
+
         return view('view_tool::web.facebook.avatar', $this->render($data));
-    }
-
-    public function postAvatar()
-    {
-        // open an image file
-        $path = request()->file('avatar')->store('public/upload/tool/facebook-avatar/' . date('Y/m/d'));
-
-        $img = Image::make(storage_path('app/' . $path));
-
-        // resize image instance
-        $img->resize(500, 500);
-
-        // insert a watermark
-        $img->insert(public_path('site/img/tich-xanh-iframe.png'), 'center');
-
-        // save image in desired format
-        $img->save(storage_path('app/public/upload/tool/facebook-avatar/' . date('Y/m/d/') . time() . '-ok-img.png'))->mime();
-
-        unlink(storage_path('app/' . $path));
-
     }
 }
